Split minimal FSDP module mixin
